+++ src/Network/Network.Test/ScenarioTests/RouteServerTests.cs
@@ -34,5 +34,13 @@ namespace Commands.Network.Test.ScenarioTests
         {
             TestRunner.RunTestScript(string.Format("Test-RouteServerPeerRoutes"));
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        [Trait(Category.Owner, NrpTeamAlias.virtualwan)]
+        public void TestRouteServerPeerWithRoutingConfiguration()
+        {
+            TestRunner.RunTestScript(string.Format("Test-RouteServerPeerWithRoutingConfiguration"));
+        }
     }
 }