+++ src/Network/Network/Models/PSVirtualNetworkGatewayConnection.cs
@@ -94,6 +94,14 @@ namespace Microsoft.Azure.Commands.Network.Models
         [Ps1Xml(Label = "GatewayCustomBgpIpAddresses", Target = ViewControl.Table)]
         public List<PSGatewayCustomBgpIpConfiguration> GatewayCustomBgpIpAddresses { get; set; }
 
+        public PSRoutingConfiguration RoutingConfiguration { get; set; }
+
+        [JsonIgnore]
+        public string RoutingConfigurationText
+        {
+            get { return JsonConvert.SerializeObject(RoutingConfiguration, Formatting.Indented, new JsonSerializerSettings() { NullValueHandling = NullValueHandling.Ignore }); }
+        }
+
         [JsonIgnore]
         public string VirtualNetworkGateway1Text
         {